feat(cli): split CLI into multiple files and refactor: - lazy load modules - add `patch` command
(WIP, apps fail)
