+++ packages/ui/rollup.config.js
@@ -10,12 +10,14 @@ const external = [
   "child_process",
   "fs/promises",
   "klona/json",
+  "module",
   "package-manager-detector/commands",
   "package-manager-detector/detect",
   "path",
   "react/jsx-runtime",
-  "tailwindcss/colors.js",
+  "readline",
   "tailwindcss/plugin",
+  "tailwindcss/version",
   ...Object.keys({
     ...packageJson.dependencies,
     ...packageJson.devDependencies,
