+++ packages/ui/src/cli/commands/build.ts
@@ -1,0 +1,7 @@
+import { generateClassList } from "./generate-class-list";
+import { setupOutputDirectory } from "./setup-output-directory";
+
+export async function build() {
+  await setupOutputDirectory();
+  await generateClassList();
+}
