+++ packages/ui/src/cli/commands/plugins/setup-plugin-rolldown.ts
@@ -1,0 +1,12 @@
+import path from "path";
+import { pluginName, pluginPath } from "../../consts";
+import { addPluginToConfig } from "../../utils/add-plugin-to-config";
+
+export async function setupPluginRolldown(configPath: string) {
+  await addPluginToConfig({
+    configKey: "plugins",
+    configPath,
+    pluginImportPath: path.join(pluginPath, "rolldown"),
+    pluginName,
+  });
+}
