+++ packages/ui/src/cli/commands/plugins/setup-plugin-tanstack-start.ts
@@ -1,0 +1,12 @@
+import path from "path";
+import { pluginName, pluginPath } from "../../consts";
+import { addPluginToConfig } from "../../utils/add-plugin-to-config";
+
+export async function setupPluginTanStackStart(configPath: string) {
+  await addPluginToConfig({
+    configKey: "vite.plugins",
+    configPath,
+    pluginImportPath: path.join(pluginPath, "vite"),
+    pluginName,
+  });
+}
