+++ packages/ui/src/cli/commands/plugins/setup-plugin-webpack.ts
@@ -1,0 +1,12 @@
+import path from "path";
+import { pluginName, pluginPath } from "../../consts";
+import { addPluginToConfig } from "../../utils/add-plugin-to-config";
+
+export async function setupPluginWebpack(configPath: string) {
+  await addPluginToConfig({
+    configKey: "plugins",
+    configPath,
+    pluginImportPath: path.join(pluginPath, "webpack"),
+    pluginName,
+  });
+}
