+++ packages/ui/src/cli/commands/setup-output-directory.ts
@@ -1,0 +1,11 @@
+import fs from "fs/promises";
+import { outputDir } from "../consts";
+
+export async function setupOutputDirectory() {
+  try {
+    await fs.access(outputDir);
+  } catch {
+    console.log(`Creating ${outputDir} directory...`);
+    await fs.mkdir(outputDir, { recursive: true });
+  }
+}
