+++ packages/ui/src/plugin/bun.ts
@@ -1,5 +1,6 @@
 import type { BunPlugin } from "bun";
-import { build, dev } from "../cli";
+import { build } from "../cli/commands/build";
+import { dev } from "../cli/commands/dev";
 import { pluginName } from "./index";
 
 export default {
