+++ packages/ui/src/plugin/esbuild.ts
@@ -1,5 +1,6 @@
 import { Plugin } from "esbuild";
-import { build, dev } from "../cli";
+import { build } from "../cli/commands/build";
+import { dev } from "../cli/commands/dev";
 import { pluginName } from "./index";
 
 export default (): Plugin => ({
