+++ packages/ui/src/plugin/modernjs.ts
@@ -1,5 +1,6 @@
 import type { AppTools, CliPlugin } from "@modern-js/app-tools";
-import { build, dev } from "../cli";
+import { build } from "../cli/commands/build";
+import { dev } from "../cli/commands/dev";
 import { pluginName } from "./index";
 
 export default (): CliPlugin<AppTools> => ({
