+++ packages/ui/src/plugin/nextjs.ts
@@ -1,5 +1,6 @@
 import type { NextConfig } from "next";
-import { build, dev } from "../cli";
+import { build } from "../cli/commands/build";
+import { dev } from "../cli/commands/dev";
 
 export default function withFlowbiteReact(nextConfig: NextConfig): NextConfig {
   if (process.env.NODE_ENV === "development") {
