+++ packages/ui/src/plugin/rspack.ts
@@ -1,5 +1,6 @@
 import type { RspackPluginInstance } from "@rspack/core";
-import { build, dev } from "../cli";
+import { build } from "../cli/commands/build";
+import { dev } from "../cli/commands/dev";
 import { pluginName } from "./index";
 
 // TODO: export default is broken
